Email_Agent_Test/email_summarizer_agent/requirements.txt
langchain
langchain-deepseek
pandas
requests
python-dotenv
cirtusai_sdk